In this commit, I have added four different models.
The first model is the basic patient model. It houses the patient's biodata.

The second model is for the "presenting complaints". Here we can add presenting complaints at reception.

The third model is "Initial tretment". It allows for entering the immediate treatment.

The fourth model is for "Investigations", to add any investigations that have been done by the A&E department.

Other commits include bundling of the gem package so the gems can be installed locally.